# swagger.yaml
openapi: 3.0.3
info:
  title: Academic & Internship Management Platform
  description: |
    Full API for managing the internship lifecycle. 
    Role-based Access Control (RBAC) is implemented:
    - **ADMINISTRATOR**: Full control over users, companies, and defense scheduling.
    - **TEACHER**: Supervision, internship status updates, and report grading.
    - **STUDENT**: Profile viewing, internship monitoring, and PDF report submission.
  version: 1.2.1

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Authentication
  - name: Administrator Only
  - name: Users
  - name: Companies
  - name: Internships
  - name: Reports & Evaluations
  - name: Defenses

paths:
  # --- AUTHENTICATION ---
  /auth/register:
    post:
      tags: [Authentication]
      summary: User registration
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User successfully registered
        '400':
          $ref: '#/components/responses/BadRequest'

  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful, returns JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        '401':
          description: Invalid credentials

  # --- USERS ---
  /users:
    get:
      tags: [Administrator Only, Users]
      summary: "[ADMIN] List all users with filters"
      operationId: getUsers
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [ADMINISTRATOR, TEACHER, STUDENT]
        - name: major
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
    post:
      tags: [Administrator Only, Users]
      summary: "[ADMIN] Create a specific user"
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User created

  /users/{email}:
    parameters:
      - name: email
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Users]
      summary: Get user details
      operationId: getUserByEmail
      responses:
        '200':
          description: User details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '404':
          description: User not found
    put:
      tags: [Administrator Only, Users]
      summary: "[ADMIN] Update user profile or role"
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '204':
          description: User updated successfully
    delete:
      tags: [Administrator Only, Users]
      summary: "[ADMIN] Remove user from system"
      operationId: deleteUser
      responses:
        '204':
          description: User deleted successfully

  /users/{email}/activate:
    parameters:
      - name: email
        in: path
        required: true
        schema:
          type: string
    patch:
      tags: [Administrator Only, Users]
      summary: "[ADMIN] Activate/Validate a user account"
      description: Enables a guest/pending account so the user can log in.
      operationId: activateUser
      security:
        - bearerAuth: []
      responses:
        '204':
          description: User activated successfully
        '403':
          description: Forbidden - Only administrators can perform this action
        '404':
          description: User not found

  # --- COMPANIES ---
  /companies:
    get:
      tags: [Companies]
      summary: List all partner companies
      operationId: getCompanies
      responses:
        '200':
          description: List of companies retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyDTO'
    post:
      tags: [Administrator Only, Companies]
      summary: "[ADMIN] Add a new partner company"
      operationId: createCompany
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyDTO'
      responses:
        '201':
          description: Company added successfully

  # --- INTERNSHIPS ---
  /internships:
    get:
      tags: [Internships]
      summary: List internships with status/student filters
      operationId: getInternships
      parameters:
        - name: studentEmail
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [ONGOING, COMPLETED, VALIDATED, REJECTED]
      responses:
        '200':
          description: List of internships retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InternshipDTO'
    post:
      tags: [Administrator Only, Internships]
      summary: "[ADMIN] Create a new internship offer"
      operationId: createInternship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternshipDTO'
      responses:
        '201':
          description: Internship created successfully

  /internships/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    put:
      tags: [Administrator Only, Internships]
      summary: "[ADMIN] Assign student and teacher to internship"
      operationId: assignUsersToInternship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternshipDTO'
      responses:
        '200':
          description: Internship updated/assigned successfully

  /internships/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    patch:
      tags: [Internships]
      summary: "[TEACHER] Update internship progress status"
      operationId: updateInternshipStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [ONGOING, COMPLETED, VALIDATED, REJECTED]
      responses:
        '200':
          description: Status updated successfully

  # --- REPORTS ---
  /internships/{id}/report:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Reports & Evaluations]
      summary: Get report metadata for an internship
      operationId: getReportByInternshipId
      responses:
        '200':
          description: Report metadata retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDTO'
    post:
      tags: [Reports & Evaluations]
      summary: "[STUDENT] Upload final report (PDF)"
      operationId: uploadReport
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Report uploaded successfully

  /reports/{reportFileName}/evaluation:
    parameters:
      - name: reportFileName
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Reports & Evaluations]
      summary: "[TEACHER] Grade and comment on report"
      operationId: evaluateReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationDTO'
      responses:
        '200':
          description: Evaluation submitted successfully

  # --- DEFENSES ---
  /defenses:
    get:
      tags: [Defenses]
      summary: List scheduled oral exams
      operationId: getDefenses
      responses:
        '200':
          description: List of defenses retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DefenseDTO'
    post:
      tags: [Administrator Only, Defenses]
      summary: "[ADMIN] Schedule a defense (Date and Room)"
      operationId: createDefense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DefenseDTO'
      responses:
        '201':
          description: Defense scheduled successfully

  /me:
    get:
      tags: [Users]
      summary: Get my own profile
      description: Returns the profile of the user identified by the JWT.
      operationId: getMyProfile
      responses:
        '200':
          description: OK - Current user profile retrieved # LA DESCRIPTION MANQUANTE ÉTAIT ICI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '401':
          description: Unauthorized - Invalid or missing token
    patch:
      tags: [Users]
      summary: Update my own profile
      operationId: updateMyProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '401':
          description: Unauthorized

components:
  schemas:
    # --- AUTH & USERS ---
    UserRegistration:
      type: object
      required: [email, password, role, lastName, firstName]
      properties:
        email: { type: string, format: email, example: "jean.dupont@eseo.fr" }
        password: { type: string, minLength: 8 }
        lastName: { type: string }
        firstName: { type: string }
        role: { type: string, enum: [ADMINISTRATOR, TEACHER, STUDENT] }
        major: { type: string, example: "Computer Science" }

    UserDTO:
      type: object
      properties:
        email: { type: string, format: email }
        lastName: { type: string }
        firstName: { type: string }
        role: { type: string }
        major: { type: string }
        active: { type: boolean }

    LoginRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string }

    AuthToken:
      type: object
      properties:
        token: { type: string }

    # --- BUSINESS LOGIC ---
    CompanyDTO:
      type: object
      required: [siret, corporateName]
      properties:
        siret: { type: string, minLength: 14, maxLength: 14, example: "12345678901234" }
        corporateName: { type: string }
        address: { type: string }
        contactEmail: { type: string, format: email }
        contactPhone: { type: string }

    InternshipDTO:
      type: object
      required: [objective, startDate, companySiret]
      properties:
        id: { type: integer, format: int64 }  
        objective: { type: string }
        startDate: { type: string, format: date }
        durationWeeks: { type: integer }
        status: { type: string, enum: [ONGOING, COMPLETED, VALIDATED, REJECTED] }
        companySiret: { type: string }
        studentEmail: { type: string }
        teacherEmail: { type: string }

    ReportDTO:
      type: object
      properties:
        id: { type: integer, format: int64 }
        fileName: { type: string }
        submissionDate: { type: string, format: date-time }
        internshipId: { type: integer, format: int64 }
        evaluation: { $ref: '#/components/schemas/EvaluationDTO' }

    EvaluationDTO:
      type: object
      properties:
        grade: { type: number, format: float, minimum: 0, maximum: 20 }
        comment: { type: string }
        date: { type: string, format: date-time }
        teacherEmail: { type: string }
        reportFileName: { type: string }

    DefenseDTO:
      type: object
      required: [date, room, studentEmail]
      properties:
        id: { type: integer, format: int64 }
        date: { type: string, format: date-time }
        room: { type: string }
        studentEmail: { type: string }

    ErrorResponse:
      type: object
      properties:
        code: { type: string }
        message: { type: string }

  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []